Fix subtle issue resolving Young Witch's Bane card
Resolves #13.

Since we were picking the bane card off the end of what we've already picked and
it has to be cost 2 or 3 we may have already consumed all 3/2s in the set of
kingdom cards. This tests to see if we have done that then tries to generate a
setup again. Not the best approach but it's kind of an edge case.